chrome/content/mochitest/tests/test_domainRestrictions.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for domain-restricted filters</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>

  <meta http-equiv="Content-Type" value="text/html; charset=utf-8"/>
  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">

  </div>
  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    prepareFilterComponents();

    function testActive(text, domain, expectedActive, expectedOnlyDomain)
    {
      let filter = Filter.fromText(text);
      is(filter.isActiveOnDomain(domain), expectedActive, text + " active on " + domain);
      is(filter.isActiveOnlyOnDomain(domain), expectedOnlyDomain, text + " only active on " + domain);
    }

    testActive("foo", null, true, false);
    testActive("foo", "com", true, false);
    testActive("foo", "example.com", true, false);
    testActive("foo", "foo.example.com", true, false);
    testActive("foo", "mple.com", true, false);

    testActive("#foo", null, true, false);
    testActive("#foo", "com", true, false);
    testActive("#foo", "example.com", true, false);
    testActive("#foo", "foo.example.com", true, false);
    testActive("#foo", "mple.com", true, false);

    testActive("foo$domain=example.com", null, false, false);
    testActive("foo$domain=example.com", "com", false, true);
    testActive("foo$domain=example.com", "example.com", true, true);
    testActive("foo$domain=example.com", "foo.example.com", true, false);
    testActive("foo$domain=example.com", "mple.com", false, false);

    testActive("example.com#foo", null, false, false);
    testActive("example.com#foo", "com", false, true);
    testActive("example.com#foo", "example.com", true, true);
    testActive("example.com#foo", "foo.example.com", true, false);
    testActive("example.com#foo", "mple.com", false, false);

    testActive("foo$domain=example.com|foo.example.com", null, false, false);
    testActive("foo$domain=example.com|foo.example.com", "com", false, true);
    testActive("foo$domain=example.com|foo.example.com", "example.com", true, true);
    testActive("foo$domain=example.com|foo.example.com", "foo.example.com", true, false);
    testActive("foo$domain=example.com|foo.example.com", "mple.com", false, false);

    testActive("example.com,foo.example.com#foo", null, false, false);
    testActive("example.com,foo.example.com#foo", "com", false, true);
    testActive("example.com,foo.example.com#foo", "example.com", true, true);
    testActive("example.com,foo.example.com#foo", "foo.example.com", true, false);
    testActive("example.com,foo.example.com#foo", "mple.com", false, false);

    testActive("foo$domain=~foo.example.com", null, true, false);
    testActive("foo$domain=~foo.example.com", "com", true, false);
    testActive("foo$domain=~foo.example.com", "example.com", true, false);
    testActive("foo$domain=~foo.example.com", "foo.example.com", false, false);
    testActive("foo$domain=~foo.example.com", "mple.com", true, false);

    testActive("~foo.example.com#foo", null, true, false);
    testActive("~foo.example.com#foo", "com", true, false);
    testActive("~foo.example.com#foo", "example.com", true, false);
    testActive("~foo.example.com#foo", "foo.example.com", false, false);
    testActive("~foo.example.com#foo", "mple.com", true, false);

    testActive("foo$domain=example.com|~foo.example.com", null, false, false);
    testActive("foo$domain=example.com|~foo.example.com", "com", false, true);
    testActive("foo$domain=example.com|~foo.example.com", "example.com", true, true);
    testActive("foo$domain=example.com|~foo.example.com", "foo.example.com", false, false);
    testActive("foo$domain=example.com|~foo.example.com", "mple.com", false, false);

    testActive("example.com,~foo.example.com#foo", null, false, false);
    testActive("example.com,~foo.example.com#foo", "com", false, true);
    testActive("example.com,~foo.example.com#foo", "example.com", true, true);
    testActive("example.com,~foo.example.com#foo", "foo.example.com", false, false);
    testActive("example.com,~foo.example.com#foo", "mple.com", false, false);

    testActive("foo$domain=example.com|~com", null, false, false);
    testActive("foo$domain=example.com|~com", "com", false, true);
    testActive("foo$domain=example.com|~com", "example.com", true, true);
    testActive("foo$domain=example.com|~com", "foo.example.com", true, false);
    testActive("foo$domain=example.com|~com", "mple.com", false, false);

    testActive("example.com,~com#foo", null, false, false);
    testActive("example.com,~com#foo", "com", false, true);
    testActive("example.com,~com#foo", "example.com", true, true);
    testActive("example.com,~com#foo", "foo.example.com", true, false);
    testActive("example.com,~com#foo", "mple.com", false, false);

    testActive("foo$domain=nnnnnnn.nnn", null, false, false);
    testActive("foo$domain=nnnnnnn.nnn", "com", false, false);
    testActive("foo$domain=nnnnnnn.nnn", "example.com", false, false);
    testActive("foo$domain=nnnnnnn.nnn", "foo.example.com", false, false);
    testActive("foo$domain=nnnnnnn.nnn", "mple.com", false, false);

    testActive("nnnnnnn.nnn#foo", null, false, false);
    testActive("nnnnnnn.nnn#foo", "com", false, false);
    testActive("nnnnnnn.nnn#foo", "example.com", false, false);
    testActive("nnnnnnn.nnn#foo", "foo.example.com", false, false);
    testActive("nnnnnnn.nnn#foo", "mple.com", false, false);
  </script>
  </pre>
</body>
</html>
